Rename `sampling_predictive`→`sampling_forward`
