+++ pymc/tests/test_sampling_forward.py
